requirements.txt
requests
beautifulsoup4
Flask
qrcode[pil]
Pillow
